+++ EssentialApp/EssentialAppTests/Helpers/XCTestCase+FeedImageDataLoader.swift
@@ -1,32 +1,34 @@
 //
-//  FeedLoader+XCTestCaseHelpers.swift
+//  XCTestCase+FeedImageDataLoader.swift
 //  EssentialAppTests
 //
 //  Created by TCode on 28/5/22.
 //
 
-import EssentialFeed
 import XCTest
+import EssentialFeed
 
-protocol FeedLoaderTestCase: XCTestCase {}
+protocol FeedImageDataLoaderTestCase: XCTestCase {}
 
-extension FeedLoaderTestCase {
-
-    func expect(_ sut: FeedLoader, toCompleteWith expectedResult: FeedLoader.Result, file: StaticString = #file, line: UInt = #line) {
+extension FeedImageDataLoaderTestCase {
+    
+    func expect(_ sut: FeedImageDataLoader, toCompleteWith expectedResult: FeedImageDataLoader.Result, when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
         let exp = expectation(description: "Wait for load completion")
 
-        sut.load { receivedResult in
+        _ = sut.loadImageData(from: anyURL()) { receivedResult in
             switch (receivedResult, expectedResult) {
             case let (.success(receivedFeed), .success(expectedFeed)):
                 XCTAssertEqual(receivedFeed, expectedFeed, file: file, line: line)
             case (.failure, .failure):
                 break
             default:
-                XCTFail("Expected \(expectedResult), got \(receivedResult) instead")
+                XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
             }
 
             exp.fulfill()
         }
+
+        action()
 
         wait(for: [exp], timeout: 1.0)
     }
